:truck: moving describeemgapi mixin to a new describe submodule
